responsive navbar fix
