+++ src/pages/HomePage.jsx
@@ -3,13 +3,13 @@ import About from "../components/landingpage/About";
 import Article from "../components/landingpage/Article";
 import Forum from "../components/landingpage/Forum";
 
-import NavbarPage from "../components/landingpage/NavbarPage";
+import HomeNavbar from "../components/landingpage/HomeNavbar";
 import Footer from "../components/Footer";
 
 const HomePage = () => {
   return (
     <>
-      <NavbarPage />
+      <HomeNavbar />
       <HeroBanner />
       <About />
       <Article />
